Differenciate installtion type for the GDK
Developers leave their GDK running sometimes, and as such these will
send usage pings like any other instance. With this change all
installations that are not running in production mode will be marked as
`gitlab-development-kit`.

One choice was not to send data at all, but now the data will provide
insight in how big this problem was and if corrections on historic data
is needed.