Spike: begin building class for migration
This commit begins a naive form of the UpdateFingerprints class,
which we will use in a migration to update the fingerprints on
occurrences and their matching feedback.

Currently, it fetches occurrences that have feedback and iterates
over each one to update its fingerprint.

It does not yet update the fingerprint of the feedback,
and it does not batch load or batch update,
both of which we will probably want.