Remove "smart" image size reset
There are complexities that arise when resetting an image's
after resizing, _if_ we try to reset it based
on the resized dimensions. If we fix it to the original image's
dimensions, it's a lot nicer to work with.

This introduces a potentially undesirable UX and will
be discussed with the team for future work.