Protect against incorrectly named files
There was an incorrect assumption that content_type_whitelist would
parse the file contents to determine mime type. Unfortunately this only
occurs in CarrierWave 2 and not 1.3.1 as we are currently using.

Currently the mime type is determined by the file extension despite the
method name of content_type_whitelist. This change will add the
CarrierWave 2 behaviour of content_type_whitelist.