Simplify award emoji mutation authorization
Places extra authorization call inside authorized!

This places the extra `emoji_awardable?` check in the
`authorized!` method of the `AuthorizeResource` concern, which
is the method responsible for determining what authorization means for a
particular mutation.

This leads to code simplification and more uniform messages in line with
the rest of the API.