Add a factory capable of making either an epic or an epic issue
This is really useful when we need to test behaviour in epic trees,
where an item might be either an epic or an epic issue. Failure to test
this assumption means we have code that is silently buggy or can fail at
runtime.

Some of these issues are discovered here and fixed (notably, reloading
the relative position requires custom code for these two classes, since
they are queried for in a union).