Update the smart virtual list API
We are in the middle of upgrading the
vue-virtual-scroll-list package version. As
part of this, we need to update this smart
virtual list wrapper component.

This will break the components that use
this shared component. This change is not
being merged into master until we update
all of the affected components.