Add tests for scoped_items and relative_siblings
The RelativePositioning system relies on `scoped_items` and
`relative_siblings` for many operations. These methods rely on some
guarantees about how they include and exclude items. These assumptions
are tested here.

These changes fix tests that are designed to test for correct behaviour
of the relative positioning sibling queries, in particular excluding
irrelevant items and the current item.

The main change here is that we add the group ID to the query, which is
important for top-level epic queries. Without this restriction, we
accidentally include any epic which has no parent from any group at all.

This change requires that all members of the union support `group_id` as
a result column. We handle this by delegating to Epic from EpicIssue,
and injecting the value directly into the select list on the epic issue
side of the queery.