Random index selection strategy
This adds a random selection strategy to rebuild indexes. We tackle 10
indexes at a time and select rather large ones randomly. This is subject
to improvement later but it'll get us going (basically removing the need
to specify an index name manually).

Also, as suggested earlier, this adds a view to conveniently access
information about indexes and introduces a corresponding ActiveRecord
model.