Track statistics for index rebuilds
This keeps track of index rebuilds and in particular about the change in
terms of ondisk size for the index. We also track execution time through
start/end timestamps.

For now, this is only used manually (e.g. to check the impact/success of
reindexing efforts).

Later, it may get used to more cleverly select indexes for rebuilding.
For example, we might prefer indexes that have had high levels of bloat
before or indexes that haven't been rebuilt recently. This can be
derived from the statistics.