Prevent 2fa token generation in the same session
If the token is already generated, we store this state so that
other redirects to this page won't regenerate the token unless
they come in a different session.

The reasoning behind it is, we have multiple resources through
the UploadController (avatars, logos) that can redirect to this
page, which means the token would be re-generated after the QR
was already generated and shown to the user, resulting in
permanently invalid 2FA codes.