Install registry CA cert in k8s runner when present
GitLab's registry can be configured with a self-signed certificate.
This is common in development environments, but may also be done in
corporate environments.

By default, self-signed certificates are treated as insecure, and
rejected by docker-in-docker builds.

To make for a more seamless Auto DevOps experience, and simplify local
development, we intorduce an optional configuration field registry.ca,
and add configuration to the GitLab-managed runner app to mount the
certificate (if present) at runtime at
/etc/docker/certs.d/$REGISTRY_HOST_PORT/ca.crt.