Fix Webpack not working over HTTPS in development
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42485 broke
Webpack manifest loading when HTTPS was used in development. The GET
request for `/assets/webpack/manifest.json` is proxied through the Rails
server, but it was previously failing because it was making an HTTP
request to an HTTPS port. We fix this by enabling HTTPS in this GET
request if GitLab Rails is using HTTPS.