Use author name snapshot on audit event
Dynamic preloading of author names from associated table is too slow for
a large number of audit events. As a result, Audit Log export to CSV is
timed out and causes disruption to users. This change ensures we use
the available author name snapshot on the audit_event record instead of
reaching out to the associated `User` table via `BatchLoader`.