Update database helpers to set the current_schema
- Update all migration and schema helpers that access
pg_class, pg_index, pg_constraing and similar catalog
yables to properly set the current schema when fetching
information about database objects.
- Update ReltuplesCountStrategy.get_statistics to use
the current schema so that we don’t accidentally fetch
the statistics for a table with the same name in a
schema different than the one GitLab uses.
- Add specs to check that migration helpers do not
generate conflicts in the presence of multiple schemas
with the same object names.