Global lease for database reindexing
The idea here is to only ever have one reindexing process running at a
time.

We generously choose the timeout for this (1 day), but in the regular
case we expect things to succeed much quicker. The timeout needs to be
above 3 times the statement timeout for reindexing (3 steps that can
take up to 6 hours currently). Overall, it doesn't hurt to set a high
timeout for this - it may just slow reindexing process down (chances are
low for this).