Fix primary write location not updating in Sidekiq
This will fix the intermittent errors we see with merge requests not
merging due to replication lag.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/42435 attempted to
fix the problem, but it did not work because EE databaase load balancing
is not enabled in Sidekiq. Attempts to write via the primary database
does not work. However, since Sidekiq always uses the primary we can
just use the main ActiveRecord connection to get the same information.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/247857