Don't fail import if repository already exists
To avoid `9:CreateRepositoryFromBundle: target directory is non-empty
continues to happen` exceptions from Gitaly, when importing a Project,
check if the project being imported already have a non-empty repository.

This is probably because the RepositoryImportWorker was enqueued twice.

Also, do not destroy the existing repository if it already exists. To
avoid deleting user data.