Mark sidekiq queues that require local shared disk access
* Queues marked with this require some form of access for data
that is shared among various components.  An example might be that the
queue writes data to a local directory that is quickly displayed to the
user.
* In scenarios where large installations exist, this requires some form
of shared stored, commonly NFS, to be configured where sidekiq can write
the data, but then a set of web front ends have the same NFS share such
that they are capable of reading that data and vice versa.
* This commit marks all queues _known_ to have such a dependency
* As this dependency is removed, this tag should be removed