Remove references to nonexistent feature flag
Apparently the g_compliance_dashboard_feature never existed. It was
used as an example in documentation and test coverage, so a YAML file
was created for it as if it was an actual feature flag. This deletes
that YAML file and replaces all references with actual feature flag
names. This should avert future confusion.