Use connectedCallback to initialise GlEmoji
Attributes may not be available in the constructor.
This causes problems, as data attributes,
like emoji name, may not be available when we need them.
Race conditions may be present which means emojis
may not render.

This commit uses the connectedCallback,
which will ensure the element has been mounted to the DOM
before we initialize the emoji.