Track unique wiki page views in Usage Ping
We currently only have a total count of wiki page views, and also a
combined unique users counter for creates/updates/deletes, which is
increased automatically through `TrackUniqueEvents` for any actions that
create `Event` records.

We want to include wiki page views in the same HLL counter, so we use
the same event name `wiki_action`.