Add Job.trace to the GraphQL query system
This adds traces as fields of jobs, at `CiJob.trace`. Raw and HTML
output is supported, as are job trace sections, and requesting
a specific tail length.

Due to the cost of fetching traces, the complexity is set very high.

An example query is:

```graphql
query($path: ID!, $name: String) {
  project(fullPath: $path) {
    pipelines(first: 1) {
      nodes {
        jobs {
          nodes {
            name
            duration
            trace { raw }
          }
        }
      }
    }
  }
}
```

We use a BuildTraceSection model for trace sections
This uses the pre-parsed build-trace sections rather than parsing the
trace each time.

Traces are authorized with read_build_trace