Create CSV Issue Imports table
We need to keep a record of issue imports that a user has performed so
that we can report on the usage of this feature. Currently, we have no
persisted record of this action in the database so we cannot
retrospectively see how often this has occurred.

Create the CsvIssueImport model which can store a reference to the user
that performed the action, and which project the issues were imported
into.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/44742
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/235732