Add a confirmation prompt to lock and unlock path locks
Many users have accidentally clicked on the "Lock" and "Unlock" buttons
when looking at a specific file in a repository. This has led to
mysterious merge errors due to the backend not surfacing the right error
message.

To reduce the chance of accidentally clicking, we now prompt a
confirmation modal before toggling the state.