Fix missing newlines in blob copy and paste in Firefox
Firefox and IE11 omit newlines when `span` blocks are used in the blob
view. `div` blocks works across all browsers, so use that to fix this
issue.

There are still a few surrounding newlines because of the other HTML
blocks (e.g. `pre` and `code`), but this change is a vast
improvement. See https://bugzilla.mozilla.org/show_bug.cgi?id=1273836
for more details.

This commit changes all references of highlighted code from `span` to
`div.` There are a lot of them.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/18534