Upgrade licensee in step with Gitaly
In order to upgrade to libgit2 v1.0 and newer, Gitaly has to bump some
of its dependencies to a newer versions which does allow use of the new
major version of libgit2. One of these Gems is licensee, which has
recently gained compatibility with Rugged v1.0 via commit 4653c1595
(gemspec: Allow use of Rugged v1.0, 2020-07-22).

This commit upgrades licensee to a version which contanis above commit.
As there is no release yet which includes it, we instead upgrade to the
latest commit on master. This is the same version Gitaly uses.