Add ability to sort issues and merge_requests: UI
When searching for Issues and Merge Request sometimes you don't want to
see the most accurate result. You might want to see the the newest
result.
Now you can change from Relevant results to newest or oldest results. As
we add more filtering in the future this will be even more valuable.