Fix query to find the closest group integration
After propagating group-level integrations, all the inheriting
integrations have the same settings and inherit_from_id set with the id
from which they are inheriting the values.

When retrieving the default integration in order to restore the
administration settings (those coming from the group or the admin
instance), we need to look for the closest group integration which does
not inherit from its parent. Basically, the closest with
inherit_from_id = nil.