Replace existing FFs with a single of OPS type
We controlled the Image Resizing feature with two separate FFs:
-`dynamic_image_resizing_requester`
-`dynamic_image_resizing_owner`

We still want to keep the ability to disable the Resizing Feature easily.
We want to have a separate Feature Flags with an ops type,
as this will fit into our FF managing schema.
Also, we want this flag to be enabled by default.