Allow other names for VSA production stage
Previously Value Stream Analytics required that the production
environment be named exactly "production" (lowercase) to be compatible
with the default stages. This relaxes that requirement so "prod" and
"gprd" are allowed, and also so capitalization is irrelevant. This will
make VSA more fully usable by a wider array of customers, and also by
GitLab itself.