Fix excessive eagerness during authorization
This introduces laziness during authorization, by performing
authorization in lazy promises, rather than as soon as is requested.

This fixes https://gitlab.com/gitlab-org/gitlab/-/issues/217767 for
object types, but for connections/arrays the story is a little more
complicated. caching_array_resolver is introduced to help deal with
that, but it comes at the cost of returning arrays rather than
relations, so is not universally suitable.