Return fallback emoji for dummy gl-emoji component
"gl-emoji" is a web component that uses getEmojiInfo
to initialize its emoji information.

"gl-emoji" calls getEmojiInfo using its own data-attribute.

However, data-attribute can be missing such as when gl-emoji
is used in in _search_bar.html.haml. In such case,
getEmojiInfo should return a fallback emoji but this behavior was
broken in !42669.