Fix workflow:rules not accessing passed-upstream and trigger variables
This is the second attempt to fix this problem.

In "Seed" section of CreatePipelineService, we run the block that
is passed when calling the main service. This block is used for adding
the variables to the pipeline from upstream and trigger requests.

In this commit, we start to run this block before evaluating
the workflow rules.

This fix is behind a FF ci_seed_block_run_before_workflow_rules.